requirements.txt
python-dotenv>=1.0
pandas>=2.0
numpy>=1.23
rich>=13
nbformat>=5.9
openai>=1.0
scikit-learn>=1.4
jinja2>=3.1 